requirements.txt
pandas 
pydantic 
python-dotenv 
openai 
shap 
scikit-learn 
matplotlib 
plotly 
seaborn